pci: Fix locating Mac driver from PCI rom
The ROM is not on BAR 1 but BAR 6, I'm not sure what that code ever
did, Qemu uses BAR 6, and I think MOL does too.

Additionally, support using a raw PEF binary without a made up
header.

Eventually we need to support f-code from the ROM but that is not
for today.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>